more proper handling of ctrl modifier
it still looks slightly tricky. we gather the keys of keydown events in a
buffer, and try to execute the command on each press. only when it's complete
(or invalid) do we clear it. when it's not complete, we wait for the next key.
if ctrl is pressed during the last key, we add it to the char we're matching,
e.g. 'ctrl-r'. we don't have multi-key commands/motions where ctrl is used as
non-last character. so we only have to pass along if ctrl is pressed now (when
the latest key was pressed), we don't have to remember it for all previous
keys.