beter highlighting of active mode
we don't use blue anymore for command mode, because firefox uses a blue outline
for the focused element. the difference was too small. now we use yellow for
the command mode (a contrasting color). and we use green for visual and red vor
visual-line mode.

we also set an outline of 2px, to override the default blue outline of focused
elements. makes the colors stand out much better