implement first bits of an ex mode, and search with / and ?, adding a new input line to the document
regular vi(m) use the bottom line of the screen as editing line, but the
browser textarea/input elements don't have that. so we add an additional input
element (single line) to the document for the input. absolutely positioned,
just below the reference input/textarea, to document.body to prevent css
selector problems.

the new input line has basic history with arrow up/down. vi-mode can be enabled
with ctrl+{, not with escape because that cancels the input line.

the search commands *#nN now also implement counts.
and they are now implemented in visual mode to, extending the selection.

for ex mode, we only handle basic ranges like "%" (all), digits (line number),
and we operate on the selection when in visual mode. if no command is present,
we move to a line number (so ":123" will go to line 123). otherwise, we only
implement the "s" command, for regexp search & replace. we pass the regexp
string to the js RegExp constructor, we're not going to implement a compatible
regular expression matcher in js. for replacements, we convert bare "&", and
"\<number>" to the js equivalents ("$&" and "$<number>"), but leave other text
intact.

for issue #8